refactor(attachments): move AttachmentRef to ironclaw_common
AttachmentRef is a plain serde DTO with no transcript-specific behavior,
and the attachment vocabulary (AttachmentKind, IncomingAttachment) already
lives in ironclaw_common. Co-locate AttachmentRef there so ironclaw_attachments
depends on the leaf common crate instead of taking an upward edge on the
ironclaw_threads transcript-contract crate to construct it.

- Move the AttachmentRef struct from ironclaw_threads::contract to
  ironclaw_common::attachment, next to IncomingAttachment, with a doc
  cross-link noting it is the durable, byte-free projection.
- ironclaw_threads re-exports AttachmentRef from common (mirrors the
  existing AttachmentKind re-export); the public path
  ironclaw_threads::AttachmentRef is unchanged, so no consumer edits.
- ironclaw_attachments now depends on ironclaw_common, not ironclaw_threads;
  land_inbound_attachments still returns Vec<AttachmentRef>.

ironclaw_attachments stays a clean leaf (filesystem + host_api + common).
