+++ crates/ironclaw_attachments/Cargo.toml
@@ -11,9 +11,9 @@ repository = "https://github.com/nearai/ironclaw"
 publish = false
 
 [dependencies]
+ironclaw_common = { path = "../ironclaw_common", version = "0.4.2" }
 ironclaw_filesystem = { path = "../ironclaw_filesystem", version = "0.1.0" }
 ironclaw_host_api = { path = "../ironclaw_host_api", version = "0.1.0" }
-ironclaw_threads = { path = "../ironclaw_threads", version = "0.1.0" }
 thiserror = "2"
 
 [dev-dependencies]
