+++ crates/ironclaw_common/src/lib.rs
@@ -16,12 +16,14 @@ mod timezone;
 mod trust_boundary;
 mod util;
 
-pub use attachment::{AttachmentKind, IncomingAttachment, normalize_mime_type};
-// The attachment-format registry is exposed as a module (`attachment_format::lookup`,
-// `attachment_format::is_supported_mime`, …) rather than flattening its
-// generically-named functions onto the crate root, where `ironclaw_common::lookup`
-// would read meaninglessly. The two types are re-exported for convenience.
-pub use attachment_format::{AttachmentFormat, ExtractorId};
+pub use attachment::{AttachmentKind, AttachmentRef, IncomingAttachment, normalize_mime_type};
+// `attachment_format` is also a `pub mod`, but the registry query functions are
+// re-exported at the crate root because the whole attachment pipeline consumes
+// them as `ironclaw_common::is_supported_mime` / `kind_for_mime` / etc.
+pub use attachment_format::{
+    AttachmentFormat, ExtractorId, accept_attribute, accept_tokens, all_formats,
+    canonical_extension, extractor_for_mime, is_supported_mime, kind_for_mime, lookup,
+};
 pub use event::{
     AppEvent, CodeExecutionFailureCategory, JobResultStatus, OnboardingStateDto, PlanStepDto,
     SelfImprovementPhase, ToolDecisionDto,
