+++ crates/ironclaw_threads/src/contract.rs
@@ -1,4 +1,4 @@
-use ironclaw_common::AttachmentKind;
+use ironclaw_common::AttachmentRef;
 use ironclaw_host_api::{AgentId, MissionId, ProjectId, TenantId, ThreadId, UserId};
 use serde::{Deserialize, Serialize};
 
@@ -41,40 +41,6 @@ impl ThreadScope {
             invocation_id: ironclaw_host_api::InvocationId::new(),
         }
     }
-}
-
-/// A reference to a single attachment carried alongside a transcript message.
-///
-/// This is a *reference*, never the bytes. Per the crate guardrails the
-/// transcript must not hold raw runtime payloads, host paths, or secrets, so an
-/// `AttachmentRef` carries only metadata plus an opaque `storage_key` (a
-/// rendered scoped path into host-side storage, not a raw host path) and the
-/// extracted/transcribed text once an extractor has run. The bytes live behind
-/// the filesystem authority that owns `storage_key`.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct AttachmentRef {
-    /// Stable identifier for this attachment within its message.
-    pub id: String,
-    /// Image / Audio / Document.
-    pub kind: AttachmentKind,
-    /// MIME type as received at the ingress boundary. Validated upstream
-    /// against the attachment format registry; stored verbatim here.
-    pub mime_type: String,
-    /// Original filename, when the source provided one.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub filename: Option<String>,
-    /// File size in bytes, when known.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub size_bytes: Option<u64>,
-    /// Opaque storage reference for the bytes (a rendered scoped path into
-    /// host-side storage, never a raw host path). `None` until the attachment
-    /// has been landed in storage.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub storage_key: Option<String>,
-    /// Extracted document text or audio transcript, once an extractor has run.
-    /// Sanitized external data; never raw bytes.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub extracted_text: Option<String>,
 }
 
 /// Safe transcript content accepted by this boundary.
@@ -535,6 +501,7 @@ pub struct UpdateThreadGoalRequest {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use ironclaw_common::AttachmentKind;
 
     fn sample_ref() -> AttachmentRef {
         AttachmentRef {
